requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cloudinary==1.38.0
Django==4.2.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.1
pymysql==1.1.0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
